UI refactoring (#616)
This PR splits the accounts page into multiple files
