+++ packages/template/src/components-page/account-settings/email-and-auth/email-and-auth-page.tsx
@@ -1,0 +1,18 @@
+import { PageLayout } from "../page-layout";
+import { EmailsSection } from "./emails-section";
+import { MfaSection } from "./mfa-section";
+import { OtpSection } from "./otp-section";
+import { PasskeySection } from "./passkey-section";
+import { PasswordSection } from "./password-section";
+
+export function EmailsAndAuthPage() {
+  return (
+    <PageLayout>
+      <EmailsSection/>
+      <PasswordSection />
+      <PasskeySection />
+      <OtpSection />
+      <MfaSection />
+    </PageLayout>
+  );
+}
