+++ packages/template/src/components-page/account-settings/page-layout.tsx
@@ -1,0 +1,7 @@
+export function PageLayout(props: { children: React.ReactNode }) {
+  return (
+    <div className='flex flex-col gap-6'>
+      {props.children}
+    </div>
+  );
+}
