+++ packages/template/src/components-page/account-settings/teams/team-api-keys-section.tsx
@@ -1,0 +1,67 @@
+import { StackAssertionError } from "@stackframe/stack-shared/dist/utils/errors";
+import { Button } from "@stackframe/stack-ui";
+import { useState } from "react";
+import { CreateApiKeyDialog, ShowApiKeyDialog } from "../../../components/api-key-dialogs";
+import { ApiKeyTable } from "../../../components/api-key-table";
+import { useUser } from "../../../lib/hooks";
+import { TeamApiKeyFirstView } from "../../../lib/stack-app/api-keys";
+import { Team } from "../../../lib/stack-app/teams";
+import { useTranslation } from "../../../lib/translations";
+import { Section } from "../section";
+
+
+export function TeamApiKeysSection(props: { team: Team }) {
+  const user = useUser({ or: 'redirect' });
+  const team = user.useTeam(props.team.id);
+
+  if (!team) {
+    throw new StackAssertionError("Team not found");
+  }
+
+  const manageApiKeysPermission = user.usePermission(props.team, '$manage_api_keys');
+  if (!manageApiKeysPermission) {
+    return null;
+  }
+
+  return <TeamApiKeysSectionInner team={props.team} />;
+}
+
+function TeamApiKeysSectionInner(props: { team: Team }) {
+  const { t } = useTranslation();
+
+  const [isNewApiKeyDialogOpen, setIsNewApiKeyDialogOpen] = useState(false);
+  const [returnedApiKey, setReturnedApiKey] = useState<TeamApiKeyFirstView | null>(null);
+
+  const apiKeys = props.team.useApiKeys();
+
+  const CreateDialog = CreateApiKeyDialog<"team">;
+  const ShowDialog = ShowApiKeyDialog<"team">;
+
+  return (
+    <>
+      <Section
+        title={t("API Keys")}
+        description={t("API keys grant programmatic access to your team.")}
+      >
+        <Button onClick={() => setIsNewApiKeyDialogOpen(true)}>
+          {t("Create API Key")}
+        </Button>
+      </Section>
+      <ApiKeyTable apiKeys={apiKeys} />
+
+      <CreateDialog
+        open={isNewApiKeyDialogOpen}
+        onOpenChange={setIsNewApiKeyDialogOpen}
+        onKeyCreated={setReturnedApiKey}
+        createApiKey={async (data) => {
+          const apiKey = await props.team.createApiKey(data);
+          return apiKey;
+        }}
+      />
+      <ShowDialog
+        apiKey={returnedApiKey}
+        onClose={() => setReturnedApiKey(null)}
+      />
+    </>
+  );
+}
